Feat: url 변환 알고리즘 추가.
