--- src/main/java/com/young/shortenerurl/model/Encoder.java
@@ -1,7 +1,0 @@
-package com.young.shortenerurl.model;
-
-public interface Encoder {
-    String encode(long index);
-
-    long decode(String param);
-}
