+++ src/main/java/com/young/shortenerurl/model/encoder/Base62Encoder.java
@@ -1,0 +1,24 @@
+package com.young.shortenerurl.model.encoder;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class Base62Encoder extends Encoder {
+    private static final int BASE62 = 62;
+    private static final String BASE62_CHAR = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+
+    @Override
+    public String encode(long index) {
+        StringBuilder sb = new StringBuilder();
+        while(index > 0) {
+            sb.append(BASE62_CHAR.charAt((int) (index % BASE62)));
+            index /= BASE62;
+        }
+
+        if (sb.length() > MAX_LENGTH){
+            throw new IllegalArgumentException("변환된 url은 8자리를 넘어갈 수 없습니다.");
+        }
+        return sb.toString();
+    }
+
+}
